add makefile for ex00 from CPP module 00

+++ CPP-Module-00/ex00/megaphone.cpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kali <kali@student.42.fr>                  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/08/04 16:24:58 by kali              #+#    #+#             */
+/*   Updated: 2021/08/04 16:27:20 by kali             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <iostream>
 #include <string>
 
@@ -10,7 +22,7 @@ int main(int ac, char **av)
         for(int i = 1; i < ac;i++)
         {
             std::string str = av[i];
-            for(int j = 0; j < str.length(); j++)
+            for(int j = 0; j < (int)str.length(); j++)
             {
                 char c = toupper(str[j]);
                 std::cout << c;
